[BE] refactor : 페이지네이션 관련 변수 상수화
