+++ backend/reviewduck/src/main/java/com/reviewduck/common/vo/PageConstant.java
@@ -1,0 +1,9 @@
+package com.reviewduck.common.vo;
+
+public class PageConstant {
+
+    public static final int DEFAULT_PAGE = 0;
+
+    public static final int DEFAULT_SIZE = 10;
+
+}
